backend/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pdfplumber
google-generativeai
python-dotenv
